ug_feedback_system/requirements.txt
Flask==3.0.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
bcrypt==4.1.2
